feat(command): support yaml file & fix some bugs (#17)

+++ index_package/index/index.py
@@ -203,6 +203,7 @@ class Index:
     return path
 
   def _update_file_with_event(self, cursor: sqlite3.Cursor, path: str, event: Event) -> tuple[Optional[str], Optional[tuple[int, str]]]:
+    cursor.execute("SELECT id, hash FROM files WHERE scope = ? AND path = ?", (event.scope, event.path,))
     row = cursor.fetchone()
     new_hash: Optional[str] = None
     origin_id_hash: Optional[tuple[int, str]] = None
@@ -277,9 +278,11 @@ class Index:
       if cursor.fetchone() is None:
         page = self._pdf_parser.page(page_hash)
         if page is not None:
-          for index in range(len(page.annotations)):
-            self._index_db.remove(f"{page.hash}/anno/{index}/content")
-            self._index_db.remove(f"{page.hash}/anno/{index}/extracted")
+          for index, anno in enumerate(page.annotations):
+            if anno.content is not None:
+              self._index_db.remove(f"{page.hash}/anno/{index}/content")
+            if anno.extracted_text is not None:
+              self._index_db.remove(f"{page.hash}/anno/{index}/extracted")
           self._index_db.remove(page.hash)
 
     self._pdf_parser.fire_file_removed(hash)
